# server/swagger.yaml
openapi: 3.0.0
info:
  title: Cephalometric Landmarks API
  version: 1.0.0
  description: API for managing cephalometric landmarks and analysis

servers:
  - url: http://localhost:5000
    description: Development server

paths:
  /api/landmarks:
    get:
      summary: Get landmarks data
      description: Retrieve the complete set of cephalometric landmarks with their coordinates
      operationId: getLandmarks
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LandmarksResponse'
        '404':
          description: Landmarks not found
        '500':
          description: Internal server error

  /api/tracing-lines:
    get:
      summary: Get tracing lines data
      description: Retrieve the complete set of tracing lines with their paths and styling
      operationId: getTracingLines
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TracingLinesResponse'
        '404':
          description: Tracing lines not found
        '500':
          description: Internal server error

components:
  schemas:
    Coordinates:
      type: object
      properties:
        x:
          type: number
          description: X coordinate of the landmark
        y:
          type: number
          description: Y coordinate of the landmark
      required:
        - x
        - y

    Point:
      type: object
      properties:
        landmark:
          type: string
          description: Name of the landmark
        coordinates:
          $ref: '#/components/schemas/Coordinates'
      required:
        - landmark
        - coordinates

    Box:
      type: object
      properties:
        left:
          type: number
          description: Left boundary of the facial area
        right:
          type: number
          description: Right boundary of the facial area
        top:
          type: number
          description: Top boundary of the facial area
        bottom:
          type: number
          description: Bottom boundary of the facial area
      required:
        - left
        - right
        - top
        - bottom

    LandmarksResponse:
      type: object
      properties:
        status:
          type: string
          description: Status of the landmarks detection
        box:
          $ref: '#/components/schemas/Box'
        message:
          type: string
          description: Response message
        points:
          type: array
          items:
            $ref: '#/components/schemas/Point'
          description: Array of landmark points with their coordinates
      required:
        - status
        - box
        - message
        - points

    TracingLine:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the tracing line
        name:
          type: string
          description: Display name of the tracing line
        pathData:
          type: string
          description: SVG path data for drawing the line
        color:
          type: string
          description: Color of the line in hex format
        strokeWidth:
          type: number
          description: Width of the line
        strokeDasharray:
          type: string
          nullable: true
          description: Optional dash pattern for the line
        transform:
          type: string
          description: SVG transform matrix for the line
      required:
        - id
        - name
        - pathData
        - color
        - strokeWidth
        - transform

    TracingLinesResponse:
      type: object
      properties:
        status:
          type: string
          description: Status of the operation
        message:
          type: string
          description: Response message
        data:
          type: array
          items:
            $ref: '#/components/schemas/TracingLine'
          description: Array of tracing lines
      required:
        - status
        - message
        - data 